[datesort] Adding datesort
Python script that recursively loops through an input folder. For each
input file, an output folder is created named 'year/month' based on the
file's modification time (e.g. 2012/02) and a symlink to the input file
is stored in this output folder.

The result is a hierarchy of year/month folders each containing links
to the files from that month.